collapse Device inside Switch
