+++ netwalk/__init__.py
@@ -1,10 +1,10 @@
 "Main file for library"
 
 from .interface import Interface
-from .switch import Switch
+from .switch import Device, Switch
 from .fabric import Fabric
 
-__all__ = ["Interface", "Switch", "Fabric"]
+__all__ = ["Interface", "Switch", "Fabric", "Device"]
 
 
 #Taken from requests library, check their documentation
